# Conflicts:
#	app/src/main/java/nodomain/freeyourgadget/gadgetbridge/util/DeviceHelper.java
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/activities/DebugActivity.java
@@ -186,7 +186,11 @@ public class DebugActivity extends AbstractGBActivity {
                 notificationSpec.body = testString;
                 notificationSpec.sender = testString;
                 notificationSpec.subject = testString;
-                notificationSpec.sourceAppId = BuildConfig.APPLICATION_ID;
+                if (notificationSpec.type != NotificationType.GENERIC_SMS) {
+                    // SMS notifications don't have a source app ID when sent by the SMSReceiver,
+                    // so let's not set it here as well for consistency
+                    notificationSpec.sourceAppId = BuildConfig.APPLICATION_ID;
+                }
                 notificationSpec.sourceName = getApplicationContext().getApplicationInfo()
                         .loadLabel(getApplicationContext().getPackageManager())
                         .toString();
@@ -478,12 +482,19 @@ public class DebugActivity extends AbstractGBActivity {
         removeDevicePreferencesButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Context context = getApplicationContext();
-                GBApplication gbApp = (GBApplication) context;
-                List<GBDevice> devices = gbApp.getDeviceManager().getSelectedDevices();
-                for(GBDevice device : devices){
-                    GBApplication.deleteDeviceSpecificSharedPrefs(device.getAddress());
-                }
+                new AlertDialog.Builder(DebugActivity.this)
+                        .setCancelable(true)
+                        .setTitle(R.string.debugactivity_confirm_remove_device_preferences_title)
+                        .setMessage(R.string.debugactivity_confirm_remove_device_preferences)
+                        .setPositiveButton(R.string.ok, (dialog, which) -> {
+                            final GBApplication gbApp = (GBApplication) getApplicationContext();
+                            final List<GBDevice> devices = gbApp.getDeviceManager().getSelectedDevices();
+                            for(final GBDevice device : devices){
+                                GBApplication.deleteDeviceSpecificSharedPrefs(device.getAddress());
+                            }
+                        })
+                        .setNegativeButton(R.string.Cancel, (dialog, which) -> {})
+                        .show();
             }
         });
